backend/requirements.txt
fastapi==0.125.0
uvicorn==0.38.0
python-dotenv==1.2.1
python-multipart==0.0.21
pypdf==6.5.0
pydantic-ai==1.39.0
